==> code/espejo/03-estimators.Rmd <==
---
title: "Estimators"
output: html_document
---

## Read in data.
```{r}
library(readr)
las_vegas <- read_csv("../../data/las-vegas-yelp.csv")[,-1]
head(las_vegas)
```

Create the intervention node.

```{r, message=FALSE, warning=FALSE}
library(dplyr)
las_vegas <- las_vegas %>% mutate(star_intervention=ifelse(stars>=3.5, 1, 0))
```

Remove any NA's.
```{r}
las_vegas <- las_vegas[complete.cases(las_vegas),]
# length(complete.cases(las_vegas)) # 3644 complete
```

I'm going to make age and review count into bins.
```{r}
# * REVIEW COUNT BINS
# reviews     <- sort(las_vegas %>% pull(review_count))
# percentiles <- sapply(seq(0, 1, 0.1), function(x) quantile(reviews, x))[-1]
# 
# review_percentiles <- rep(NULL, times=length(reviews))
# for (ix in 1:length(percentiles)) {
#   this_per <- rev(percentiles)[ix]
#   these_ix <- which(reviews<this_per)
#   review_percentiles[these_ix] <- 11-ix
# }
# 
# review_bins <- cbind(review_count=reviews, review_percentiles=review_percentiles)
# las_vegas <- merge(las_vegas, data.frame(review_bins), by="review_count")
```

```{r}
# * REVIEW AGE BINS
# ages        <- sort(las_vegas %>% pull(age))
# percentiles <- sapply(seq(0, 1, 0.1), function(x) quantile(ages, x))[-1]
# 
# age_percentiles <- rep(NULL, times=length(reviews))
# for (ix in 1:length(percentiles)) {
#   this_per <- rev(percentiles)[ix]
#   these_ix <- which(ages<this_per)
#   age_percentiles[these_ix] <- 11-ix
# }
# 
# age_bins <- cbind(age=ages, age_percentiles=age_percentiles)
# las_vegas <- merge(las_vegas, data.frame(age_bins), by="age")
```


```{r}
las_vegas <- las_vegas %>% mutate(age_binary=age>median(age),
                                  review_binary=review_count>median(review_count))
```

## Simple substitution estimator.
I used R Lab 2.
```{r}
e_a1 <- las_vegas %>% filter(star_intervention==1) %>%
  group_by(neighborhood, american, chain, age_binary, review_binary) %>%
  summarize(e.ya1=mean(open_2019))
# nrow(e_a1)

e_a0 <- las_vegas %>% filter(star_intervention==0) %>%
  group_by(neighborhood, american, chain, age_binary, review_binary) %>%
  summarize(e.ya0=mean(open_2019))
# nrow(e_a0)

probs <- las_vegas %>% 
  group_by(neighborhood, american, chain) %>% 
  summarize(p=n()/nrow(las_vegas))

expectations <- merge(merge(e_a1, e_a0), probs)
expectations <- expectations %>% mutate(weighted.avg=(e.ya1-e.ya0)*p)
psi.p0 <- sum(expectations$weighted.avg)

psi.p0*100
```

## IPTW.
```{r}
library(nnet)
gAW_reg  <- multinom(star_intervention~chain+american+review_binary+age_binary, data=las_vegas)
gAW_pred <- predict(gAW_reg, type="probs")

gAW <- as.numeric(gAW_pred)
wt  <- gAW^-1

IPTW <- mean(wt*as.numeric(las_vegas$star_intervention==1)*las_vegas$open_2019) -
        mean(wt*as.numeric(las_vegas$star_intervention==0)*las_vegas$open_2019)
IPTW
```

```{r}
IPTW_or <- mean(wt*as.numeric(las_vegas$star_intervention==1)*las_vegas$open_2019) / mean( wt*as.numeric(las_vegas$star_intervention==1)) - mean(wt*as.numeric(las_vegas$star_intervention==0)*las_vegas$open_2019) / mean( wt*as.numeric(las_vegas$star_intervention==0))
IPTW_or
```


## TMLE.
```{r}
library(SuperLearner)
# specify the library
# install.packages("gam")
SL.library<- c("SL.glm", "SL.step", "SL.gam")

X <- X1 <- X0 <- las_vegas %>% select(star_intervention, american, review_count, age)

X1$star_intervention <- 1         # under exposure
X0$star_intervention <- 0 

QbarSL <- SuperLearner(Y=las_vegas$open_2019, 
                       X=X, 
                       SL.library=SL.library, family="binomial")

# names(las_vegas)
QbarSL


QbarAW <- predict(QbarSL, newdata=las_vegas %>% select(star_intervention, american, review_count, age))$pred

Qbar1W <- predict(QbarSL, newdata=X1)$pred
# expected injury severity, given A=0 and covariates
Qbar0W <- predict(QbarSL, newdata=X0)$pred

# Simple sub
PsiHat.SS<-mean(Qbar1W - Qbar0W)
PsiHat.SS


gHatSL<- SuperLearner(Y=las_vegas$star_intervention, X=X %>% select(-star_intervention, -business_id, -neighborhood, -name, -address, -postal_code, -stars, -categories, -open_2019)),
                       SL.library=SL.library, family="binomial")
```
